Fixed UART transmit issue; added copious comments.
Main fix was to the transmit section of the code in the function
putch(). Hadn't realised that LBUSY is 1 when the UART is busy compared
to UDREn on a normal mega MCU with a normal UART where 0 indicates busy.

Added lots of comments, as is my way, to help the poor souls that
wander this way in the future....