Fixed section addresses to actually put boot loader in boot space!
Didn't realise that the AVR used 16-bit instructions (what an
admission!) and byte-wise memory addressing. Consequently, I'd
but the boot loader in the middle of memory instead of the end of
memory. Consequently, the boot loader couldn't write/erase the
space that it was actually located in.

Three days well spent, me thinks...